Update devkit and 2 new weapon sets
Aquilonian and Brythunian weapon sets added.

New devkit addressing issues with ModsShared and the older list system for ignoring overlapping build pieces
Recompiling due to an abundance of caution.  Both components updated are actually in use in this mod, although I've not had the trouble it fixes.  The overlapping build is there simply because I copied what Funcom had.

Also, marking as v1.0 - this isn't what I'd call out of beta for something I'd work on in my job, but it's quite frankly about as good as it's going to get for me doing CE mods.