patching
much tidy up and adjusting blueprints for the 2 exploding benches..  Checks and such that existed for exploding on crafting have been added into the chaces of exploding from bench damage.
The spurious flame was the left over flame that lingers after and explosion.  The visuals for the explosion should have been children of the "Destructable" item, but the visibility wasn't following for children, just the parent item.  So the flame etc was always there.  In the end I fixed that by just turning off and on the various effects, instead of relying on the child/parent arrangement.
Fixed the benches only being removed at the tidy up stage after all explosive after effects are done.
Working on but not quite achieving making the leacher/crystalizer running frame follow whterh it's running or not.