Adding DD API
incorporating DD's accessory API primarily to not conflict with their slot locations, and to investigate "restraints" in relation to stumps etc.
This is a backend change and no new items or functionality involved.  The quiped location of some accessory items will change on the next time they are equiped, again, primarily so as to not  overlap with accessories from DD mod.