Hats and such
Shrunk Cherenkov effect on Radium torch and reduced it's time lingering.
Added a new headpiece (seems insufficient to call it a hat).  A take on the late medieval Hennin.  4 dye channels although it may take some looking to see the pearl/4th channel, they're essentially dots.
Added a test accessory ribbon.  It's tied to the pelvis bone so needs shifting with a mod like ToT! Custom, where it could be liberally plonked around.  The devkit borks at adding without rigging, and then fails to show apex physics on such a thing, so I am very very much guessing at whether this will work.  But, costs 1 silk, and is part of the everyone feat.  If it doesn't work, meh.